fix: naming of test class to match method-per-media-type support
