+++ client/integration-tests/method-per-media-type/src/test/java/io/quarkiverse/openapi/generator/it/ImagesApiSingleMethodWithMultipleMediaTypesTest.java
@@ -4,10 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import java.io.File;
 import java.lang.reflect.Method;
-import java.util.Arrays;
 import java.util.List;
-import java.util.Map;
-import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 import jakarta.ws.rs.Consumes;
@@ -22,13 +19,13 @@ import io.quarkus.test.junit.QuarkusTest;
 public class ImagesApiSingleMethodWithMultipleMediaTypesTest {
     static Stream<Class<?>> imagesApiClasses() {
         return Stream.of(
-                org.acme.json.multiple.media.types.enabled.api.ImagesApi.class,
-                org.acme.yaml.multiple.media.types.enabled.api.ImagesApi.class);
+                org.acme.json.multiple.media.types.disabled.api.ImagesApi.class,
+                org.acme.yaml.multiple.media.types.disabled.api.ImagesApi.class);
     }
 
     @ParameterizedTest
     @MethodSource("imagesApiClasses")
-    void imagesApiIsBeingGeneratedWithMethodPerMediaType(Class<?> imagesApiClass) throws NoSuchMethodException {
+    void imagesApiIsGeneratedWithSingleUpdateMethod(Class<?> imagesApiClass) throws NoSuchMethodException {
         assertImagesApiPerClass(imagesApiClass);
     }
 
@@ -39,49 +36,13 @@ public class ImagesApiSingleMethodWithMultipleMediaTypesTest {
         Method listImages = imagesApiClass.getMethod("listImages", Integer.class, Integer.class);
         assertThat(listImages.getReturnType()).isEqualTo(List.class);
 
-        assertThat(listImages.getGenericReturnType()
-                .getTypeName())
-                .contains("java.util.List");
+        Method updateImage = imagesApiClass.getMethod("updateImage", String.class, File.class);
+        assertThat(updateImage.getReturnType()).isEqualTo(Response.class);
 
-        Map<String, String> expectedMethodSuffixPerMediaType = Map.of(
-                "ImageJpeg", "image/jpeg",
-                "ImagePng", "image/png",
-                "ImageGif", "image/gif");
-
-        List<Method> updateMethods = Arrays.stream(imagesApiClass.getMethods())
-                .filter(m -> m.getName()
-                        .startsWith("updateImage"))
-                .toList();
-
-        assertThat(updateMethods.stream()
-                .map(Method::getName)
-                .collect(Collectors.toSet()))
-                .containsExactlyInAnyOrder("updateImageImageJpeg", "updateImageImagePng", "updateImageImageGif");
-
-        for (Method m : updateMethods) {
-            assertThat(m.getReturnType()).isEqualTo(Response.class);
-            assertThat(m.getParameterTypes()).containsExactly(String.class, File.class);
-
-            Consumes consumes = m.getAnnotation(Consumes.class);
-            assertThat(consumes)
-                    .as("@Consumes annotation must be present on %s", m.getName())
-                    .isNotNull();
-            assertThat(consumes.value())
-                    .as("@Consumes annotation must have exactly one media type on %s", m.getName())
-                    .hasSize(1);
-
-            String mediaType = consumes.value()[0];
-
-            String suffix = m.getName()
-                    .substring("updateImage".length());
-
-            assertThat(expectedMethodSuffixPerMediaType)
-                    .as("Unexpected suffix on %s", m.getName())
-                    .containsKey(suffix);
-
-            assertThat(mediaType)
-                    .as("Consumes media type must match method suffix on %s", m.getName())
-                    .isEqualTo(expectedMethodSuffixPerMediaType.get(suffix));
-        }
+        Consumes consumes = updateImage.getAnnotation(Consumes.class);
+        assertThat(consumes).as("@Consumes must be present on updateImage")
+                .isNotNull();
+        assertThat(consumes.value())
+                .containsExactlyInAnyOrder("image/jpeg", "image/png", "image/gif");
     }
 }
